Handle pre-populated phone number with incorrect international dial code 	* if _updateFlagFromInputVal() doesn't set initial flag, continue setting default/preferred flag
This bug occured in our app, where the user inputs an invalid number and hits "submit".  Our server
then re-renders the same page with the form pre-populated with the user's previous incorrect input.

Since no dial code is found, no '.active' flag is set, and the plugin gets in a bad state, subsequently
setting off a js error when the user tries to click on any flag.

This can be reproduced by pre-populating the input with my phone number 81770##### (w/ no country code).