Change behaviour of autoHideDialCode option
It still clears the input on blur/submit if it just contains a dial code, but now it no longer re-adds it on focus. This is due to a bug (https://github.com/jackocnr/intl-tel-input/issues/847): when submitting the form with an empty (but required) input, the browser displayed an error on the input and also focusing it, which caused the plugin to automatically add the dial code which then triggered the browser to remove the error again instantly, so it could actually never be seen. I have also just come to believe more generally that injecting a value on focus is unexpected and dangerous, and so should be avoided.