Fix issue with area codes and selected flag
Previously, if you type Kazak area code +733 you get Kazak flag, but if you change number to +734 the flag stayed as Kazak when actually it should have changed to the "root" country for that dial code (+7) because +734 is not a Kazak area code, nor a substring of one either. This change adds area code substrings to the country code map, so if you select an "area code" country like Kazak, then while you're typing the number the flag will stay as Kazak, but if you start typing a number that is not a substring of a Kazak area code, then it will default to the root country, in this case Russia.