Avoid referencing env-specific global variables at the top level
If the script is loaded outside of a browser, e.g. in a rendering pipeline that
executes the same JavaScript code on both servers and in browsers, then
references to browser-only global variables may fail.  Merely importing the
script could result in error.  In such scenarios, which certainly can and do
employ strategies for loading scripts such as the synchronous one specified for
ECMAScript’s official module system, it would be much more convenient for users
of the script to be able to import it and conditionally use the imported code,
rather than deal with the challenge of dynamically importing the script, just to
compensate for the present assumptions it makes.

Therefore, instead of referencing “window” at the top level of the module, guard
such references with typeof checks.