Ensure `loadUtils()` errors don't go unhandled
This makes sure any errors/rejections that occur when trying to load the utilities script asynchronously can be handled by developers of this package. Previously, you'd wind up with "unhandled promise rejection" errors in your console no matter what you did. The previous behavior also prevented testing async loading in Jest, which considers these async errors to be test failures.

This also makes sure that the actual error gets passed through to the `itiInstance.promise` property so developers can get access to it, and better understand what the issue was. Previously, you'd just get the promise rejected with `undefined`.

Finally, this changes the error message in builds that have the utilties built in. The previous method of producing an error caused *other* things in the library to fail (or really, never complete) and was not especially informative. There is now a more detailed message that shows up as an actual failure in all the right places instead of leaving some stuff hanging.