Add support for `utilsScript` option as a function
The argument to `loadUtils` and the value of the `utilsScript` option can now be a function that returns a promise for the utils module object instead of just a string to load them from. This makes it easier to build separate bundles for utils when using bundles like Webpack, Parcel, Vite, etc. as well as providing more flexible options for mocking or loading alternative versions of utils.

Fixes #1816.